+++ src/evrim/types/collection_list_response.py
@@ -1,10 +1,18 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-from typing import List
-from typing_extensions import TypeAlias
+from typing import List, Optional
 
+from .._models import BaseModel
 from .collection import Collection
 
 __all__ = ["CollectionListResponse"]
 
-CollectionListResponse: TypeAlias = List[Collection]
+
+class CollectionListResponse(BaseModel):
+    count: int
+
+    results: List[Collection]
+
+    next: Optional[str] = None
+
+    previous: Optional[str] = None