+++ src/evrim/types/prompt_template_list_response.py
@@ -1,10 +1,18 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-from typing import List
-from typing_extensions import TypeAlias
+from typing import List, Optional
 
+from .._models import BaseModel
 from .prompt_template import PromptTemplate
 
 __all__ = ["PromptTemplateListResponse"]
 
-PromptTemplateListResponse: TypeAlias = List[PromptTemplate]
+
+class PromptTemplateListResponse(BaseModel):
+    count: int
+
+    results: List[PromptTemplate]
+
+    next: Optional[str] = None
+
+    previous: Optional[str] = None